backend/src/evaluation/requirements.txt
anthropic==0.39.0
psycopg2-binary==2.9.9
boto3==1.34.44
numpy==1.24.3
